[AIRFLOW-6310] Group CLI commands in subcommands
