+++ airflow/cli/commands/config_command.py
@@ -16,6 +16,8 @@
 # under the License.
 """Config sub-commands"""
 from airflow.configuration import conf
+from airflow.models import Connection, Variable
+from airflow.utils import cli as cli_utils, db
 
 
 def show_config(args):
@@ -27,3 +29,14 @@ def show_config(args):
         for parameter_key, value in sorted(parameters.items()):
             print(f"{parameter_key}={value}")
         print()
+
+
+@cli_utils.action_logging
+def rotate_fernet_key(args):
+    """Rotates all encrypted connection credentials and variables"""
+    with db.create_session() as session:
+        for conn in session.query(Connection).filter(
+                Connection.is_encrypted | Connection.is_extra_encrypted):
+            conn.rotate_fernet_key()
+        for var in session.query(Variable).filter(Variable.is_encrypted):
+            var.rotate_fernet_key()
