+++ airflow/cli/commands/run_command/scheduler_command.py
