+++ tests/cli/commands/run_command/test_webserver_command.py
@@ -26,8 +26,8 @@ import psutil
 
 from airflow import settings
 from airflow.bin import cli
-from airflow.cli.commands import webserver_command
-from airflow.cli.commands.webserver_command import get_num_ready_workers_running
+from airflow.cli.commands.run_command import webserver_command
+from airflow.cli.commands.run_command.webserver_command import get_num_ready_workers_running
 from airflow.models import DagBag
 from airflow.utils.cli import setup_locations
 from tests.test_utils.config import conf_vars
